+++ src/test/java/nextstep/step1/fp/CarTest.java
@@ -1,4 +1,4 @@
-package nextstep.fp;
+package nextstep.step1.fp;
 
 import org.junit.jupiter.api.Test;
