+++ src/test/java/nextstep/step1/optional/UserTest.java
@@ -1,9 +1,9 @@
-package nextstep.optional;
+package nextstep.step1.optional;
 
 import org.junit.jupiter.api.Test;
 
-import static nextstep.optional.User.ageIsInRange1;
-import static nextstep.optional.User.ageIsInRange2;
+import static nextstep.step1.optional.User.ageIsInRange1;
+import static nextstep.step1.optional.User.ageIsInRange2;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class UserTest {
